wip: implementing multiple subqueries
Rather than have just one subquery, multiple will be supported,
separated by a | in the query syntax. This does lead to a combinatorial
explosion, especially if deeply nested and means that not all variables
will be found in the results at all times. It depends on the branch that
is being executed.

Reference: https://github.com/annotation/stam/issues/28